SO 4404-0027.  Move the code under src.
Add data files bible-be.txt (the Bible in Basic English, with one line
of text per verse of the bible, and no book headings, and just one file
heading line), and linux.words (downloaded from
http://www.math.wichita.edu/~axmann/Math714/Misc/linux.words).

Remove a number of intermediate data files of no lasting significance.
Add the README.md file to give some documentation.  Other clean-up.
