--- so-4404-0027/.gitignore
@@ -1,10 +1,0 @@
-*.dat
-*.off
-article.txt
-bible12.txt
-great.panjandrum
-misspelling.txt
-misspellings.20*.*
-new.dictionary
-sc19
-sc61
