--- so-4404-0027/article_0.txt
@@ -1,13 +1,0 @@
-I will give you some advice about lofe.
-
-Eatt more roughage;
-
-Do more than others expect you to do and do it pains;
-
-Remember what lofe tells you;
-
-do not take to heart every thing you hear.
-
-do not spend all thatt you have.
-
-do not sleep as long as you want.
