--- so-4404-0027/dictionary.txt
@@ -1,249 +1,0 @@
-a
-aaron
-about
-accomplishment
-achieve
-advantage
-advice
-again
-Age
-alive
-all
-am
-and
-answer
-appearance
-appetizers
-appreciation
-ardently
-are
-around
-as
-asks
-at
-atmosphere
-attained
-attention
-be
-because
-been
-behavior
-best
-bless
-books
-break
-but
-by
-call
-can
-change
-charitable
-chatting
-close
-closing
-comment
-complete
-completely
-continue
-cooking
-deeds
-deepest
-deeply
-demands
-dignity
-dish
-dispute
-do
-door
-dreams
-earn
-earth
-eat
-effort
-enjoy
-even
-every
-expect
-eyes
-fail
-fall
-family
-feel
-figure
-find
-first
-flow
-fool
-for
-forget
-friend
-friendly
-friendship
-gain
-get
-give
-go
-god
-golden
-great
-harmonizing
-harmony
-harvest
-have
-hear
-heart
-help
-her
-hold
-honest
-honestly
-i
-if
-in
-into
-is
-it
-judge
-kisses
-know
-knowledge
-learn
-less
-lesson
-let
-life
-likes
-little
-live
-ll
-lock
-long
-look
-looking
-love
-lover
-luck
-main
-make
-man
-marry
-may
-meaning
-money
-more
-most
-mother
-much
-must
-nature
-neglect
-never
-noble
-not
-of
-old
-on
-onto
-other
-others
-our
-out
-over
-pain
-pains
-past
-pay
-people
-person
-phone
-pick
-place
-please
-quarrel
-question
-quibble
-quick
-quickly
-rather
-read
-reasonably
-remember
-reparation
-respect
-respects
-reviving
-risk
-roughage
-rules
-s
-say
-settle
-share
-should
-shoulder
-sight
-silence
-simle
-sleep
-slowly
-smil
-smoothly
-sneezes
-some
-someone
-sorry
-speak
-spend
-stand
-success
-supreme
-take
-talk
-talking
-target
-television
-tells
-than
-that
-the
-their
-them
-there
-thing
-think
-this
-three
-time
-timeless
-times
-to
-tradition
-treat
-trust
-try
-unconditionally
-understand
-up
-valuable
-ve
-want
-wanted
-watch
-way
-what
-when
-whenever
-whevever
-while
-who
-why
-will
-with
-without
-words
-wrongdoing
-year
-yesterday
-you
-your
-yourself
