--- so-4404-0027/first_ten_lines_of_expected_output.txt
@@ -1,10 +1,0 @@
-abarim 605514 635584 666445 666728 666798 824011 2780448 3120930
-abda 1322529 1917654		
-abdeel 2852183		
-abdi 1584332 1785608 1868218		
-abiah 1582807		
-abiasaph 225550		
-abida 86292 1561385		
-abidan 504071 511413 540382 541008 554760		
-abiel 1079957 1109735 1295622 1609561		
-abiezrite 956716
