--- so-4404-0027/makefile
@@ -1,19 +1,0 @@
-# To suppress JL-specific options:
-# LDLIBS= LDFLAGS= IFLAGS=
-
-include ../etc/soq-head.mk
-
-PROG1 = sc19
-PROG2 = sc61
-PROG3 = sc59
-
-PROGS = ${PROG1} ${PROG2} ${PROG3}
-
-BASEDIR = ..
-
-all: ${PROGS}
-
-clean:
-	${RM_FR} ${DEBRIS} *.dSYM
-
-realclean:
