--- so-4404-0027/misspelling_0.txt
@@ -1,5 +1,0 @@
-eatt 41
-
-lofe 34 131
-
-thatt 209
