--- so-4404-0027/misspelling_1.txt
@@ -1,12 +1,0 @@
-bible 1769108595
-bill 43
-domain 29
-feeding 70
-from 112
-his 78
-james 1936289390
-king 0
-mcginnis 48
-ministries 57
-public 6778473
-sheep 82
