+++ src/makefile
@@ -146,6 +146,7 @@ SO_DIRECTORIES = \
 	so-4319-0158 \
 	so-4380-8114 \
 	so-4390-2898 \
+	so-4404-0027 \
 
 BUILD_DIRECTORIES = \
 	AES \
