+++ src/so-4404-0027/.gitignore
@@ -1,0 +1,13 @@
+*.dat
+*.off
+article.txt
+bible12.txt
+great.panjandrum
+misspelling.txt
+misspellings.20*.*
+mknopunct
+new.dictionary
+no-punctuation-linux.words
+sc19
+sc61
+spell-checker-ac
