+++ src/so-4404-0027/bible28.txt
@@ -1,0 +1,1 @@
+../src/data-files/bible-be.txt
