+++ src/so-4404-0027/linux.words
@@ -1,0 +1,1 @@
+../src/data-files/linux.words
